Pass knex instance to applyFilter
This fixes a regression introduced in #786.
When using a SQLite database, items could not be updated any more.
This is due to knex by default only allowing a single connection for
SQLite databases and the item update operation using applyFilter inside
a transaction which tries to use a different knex instance to connect to
the database.