app: fix sort creating items (#5484)
* app: fix sort creating items
Problem: When sorting, the relations were assigned to
another fresh created items on the related collection

Why: It was missing the primary key on nested relation.
When the primary key is missing, new item is created.

How to fix: Iterate over all fields to request and when nested field
is found, append the nested primary key.

* Make function name more descriptive, add comment, fix linter warning

Co-authored-by: rijkvanzanten <rijkvanzanten@me.com>